requirements.txt
psycopg2-binary>=2.9.7
kafka-python>=2.0.2
python-dotenv>=1.0.0
requests>=2.31.0
confluent-kafka>=2.2.0
pyarrow
msfabricpysdkcore
azure-storage-file-datalake
pandas